requirements.txt
allure-pytest==2.13.1
black==23.3.0
cryptography==40.0.1
glom==23.3.0
kubernetes==26.1.0
pytest-describe==2.1.0
pytest-spec==3.2.0
pytest==7.3.0
ruff==0.0.261
yamllint==1.31.0
